@@ -5,7 +5,8 @@ import at.kocmana.testservices.customerservice.customernetwork.model.dto.Custome
 import at.kocmana.testservices.customerservice.customernetwork.model.dto.CustomerNetworkDto;
 import at.kocmana.testservices.customerservice.customernetwork.model.mapper.CustomerRelationshipMapper;
 import java.util.List;
-import javax.validation.Valid;
+
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
