@@ -6,8 +6,9 @@ import at.kocmana.testservices.ecommerceservice.price.model.dto.PriceDto;
 import at.kocmana.testservices.ecommerceservice.price.model.mapper.PriceMapper;
 import java.time.LocalDateTime;
 import java.util.List;
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
+
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
