@@ -1,13 +1,13 @@
 package at.kocmana.testservices.ecommerceservice.purchase;
 
 import at.kocmana.testservice.commons.delay.annotation.NormallyDistributedEndpointDelaySimulation;
-import at.kocmana.testservices.ecommerceservice.purchase.model.domain.Purchase;
 import at.kocmana.testservices.ecommerceservice.purchase.model.dto.PurchaseCreationRequest;
 import at.kocmana.testservices.ecommerceservice.purchase.model.dto.PurchaseResponse;
 import at.kocmana.testservices.ecommerceservice.purchase.model.mapper.PurchaseMapper;
 import java.util.List;
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
+
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
