@@ -1,12 +1,10 @@
 package at.kocmana.testservices.productservice.config;
 
-import java.security.SecureRandom;
-import java.util.List;
-import javax.sql.DataSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
+import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -14,62 +12,58 @@ import org.springframework.security.provisioning.JdbcUserDetailsManager;
 import org.springframework.security.provisioning.UserDetailsManager;
 import org.springframework.security.web.SecurityFilterChain;
 
+import javax.sql.DataSource;
+import java.security.SecureRandom;
+import java.util.List;
+
 @Configuration
 public class WebSecurityConfiguration {
 
-  private static final String ADMIN_ROLE = "ADMIN";
-  private static final String USER_ROLE = "USER";
+    private static final String ADMIN_ROLE = "ADMIN";
+    private static final String USER_ROLE = "USER";
 
-  private final BasicAuthEntryPoint authenticationEntryPoint;
-  private final SecurityWhitelistProperties securityWhitelistProperties;
-  private final DataSource dataSource;
+    private final SecurityWhitelistProperties securityWhitelistProperties;
+    private final DataSource dataSource;
 
-  @Autowired
-  public WebSecurityConfiguration(BasicAuthEntryPoint authenticationEntryPoint,
-                                  SecurityWhitelistProperties securityWhitelistProperties, DataSource dataSource) {
-    this.authenticationEntryPoint = authenticationEntryPoint;
-    this.securityWhitelistProperties = securityWhitelistProperties;
-    this.dataSource = dataSource;
-  }
+    @Autowired
+    public WebSecurityConfiguration(SecurityWhitelistProperties securityWhitelistProperties, DataSource dataSource) {
+        this.securityWhitelistProperties = securityWhitelistProperties;
+        this.dataSource = dataSource;
+    }
 
-  @Bean
-  public UserDetailsManager users() {
-    return new JdbcUserDetailsManager(dataSource);
-  }
+    @Bean
+    public UserDetailsManager users() {
+        return new JdbcUserDetailsManager(dataSource);
+    }
 
-  @Bean
-  public SecurityFilterChain filterChain(HttpSecurity httpSecurity) throws Exception {
-    httpSecurity.authorizeRequests()
-        .antMatchers(HttpMethod.DELETE, "/**").hasRole(ADMIN_ROLE)
-        .antMatchers(HttpMethod.POST, "/product/**").hasRole(ADMIN_ROLE)
-        .antMatchers(HttpMethod.PATCH, "/product/**").hasRole(ADMIN_ROLE)
-        .antMatchers("/review/**").hasAnyRole(ADMIN_ROLE, USER_ROLE)
-        .antMatchers(HttpMethod.GET, "/**").hasAnyRole(ADMIN_ROLE, USER_ROLE)
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity httpSecurity) throws Exception {
+        httpSecurity.authorizeHttpRequests(requests -> requests
+                        .requestMatchers(HttpMethod.DELETE, "/**").hasRole(ADMIN_ROLE)
+                        .requestMatchers(HttpMethod.POST, "/product/**").hasRole(ADMIN_ROLE)
+                        .requestMatchers(HttpMethod.PATCH, "/product/**").hasRole(ADMIN_ROLE)
+                        .requestMatchers("/review/**").hasAnyRole(ADMIN_ROLE, USER_ROLE)
+                        .requestMatchers(HttpMethod.GET, "/**").hasAnyRole(ADMIN_ROLE, USER_ROLE)
 
-        .antMatchers("/actuator/**").permitAll()
-        .antMatchers(generateConfigurationWhitelist()).permitAll()
+                        .requestMatchers("/actuator/**").permitAll()
+                        .requestMatchers(generateConfigurationWhitelist()).permitAll()
 
-        .anyRequest()
-        .authenticated()
+                        .anyRequest()
+                        .authenticated())
+                .httpBasic(Customizer.withDefaults())
+                .csrf(Customizer.withDefaults());
 
-        .and()
-        .httpBasic()
-        .authenticationEntryPoint(authenticationEntryPoint);
+        return httpSecurity.build();
+    }
 
-    httpSecurity.csrf().disable()
-        .headers().frameOptions().disable();
+    private String[] generateConfigurationWhitelist() {
+        List<String> whitelist = securityWhitelistProperties.whitelist();
+        return whitelist.toArray(new String[0]);
+    }
 
-    return httpSecurity.build();
-  }
-
-  private String[] generateConfigurationWhitelist() {
-    List<String> whitelist = securityWhitelistProperties.getWhitelist();
-    return whitelist.toArray(new String[0]);
-  }
-
-  @Bean
-  public PasswordEncoder passwordEncoder() {
-    return new BCryptPasswordEncoder(10, new SecureRandom());
-  }
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder(10, new SecureRandom());
+    }
 
 }
