@@ -9,8 +9,9 @@ import at.kocmana.testservices.productservice.productinformation.model.dto.Produ
 import at.kocmana.testservices.productservice.productinformation.model.dto.ProductUpdateRequest;
 import at.kocmana.testservices.productservice.productinformation.model.mapper.ProductMapper;
 import java.util.List;
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
+
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
