@@ -12,8 +12,9 @@ import at.kocmana.testservices.productservice.productreview.model.dto.ProductRev
 import at.kocmana.testservices.productservice.productreview.model.mapper.ProductReviewMapper;
 import java.util.AbstractMap;
 import java.util.List;
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
+
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
