@@ -1,12 +1,12 @@
 package at.kocmana.testservice.commons.delay.interceptor;
 
 import at.kocmana.testservice.commons.delay.annotation.ProbabilisticEndpointDelaySimulation;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 @Component
 @Slf4j
