@@ -3,14 +3,13 @@ package at.kocmana.testservices.customerservice.controlleradvice;
 import at.kocmana.testservice.commons.responses.ErrorResponse;
 import at.kocmana.testservice.commons.responses.ValidationErrorResponse;
 import at.kocmana.testservices.customerservice.customerinformation.model.CustomerNotFoundException;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import javax.servlet.http.HttpServletRequest;
 
 import static org.springframework.http.HttpStatus.*;
 
