@@ -1,10 +1,9 @@
 package at.kocmana.testservices.productservice.productinformation.model.domain;
 
+import jakarta.persistence.Embeddable;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import javax.persistence.Embeddable;
 
 @Embeddable
 @Getter
