@@ -1,14 +1,14 @@
 package at.kocmana.testservices.productservice.productinformation.model.dto;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.Value;
 import org.hibernate.validator.constraints.Length;
-
-import javax.validation.constraints.NotBlank;
 
 @Value
 public class ProductCreationRequest {
 
-  @NotBlank @Length(max = 255)
+  @NotBlank
+  @Length(max = 255)
   String name;
   @NotBlank @Length(max = 500)
   String description;
