@@ -1,18 +1,19 @@
 package at.kocmana.testservices.productservice.productinformation.model.dto;
 
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.Value;
 import org.hibernate.validator.constraints.Length;
-
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 
 @Value
 public class ProductUpdateRequest {
 
-  @NotNull @Min(1)
+  @NotNull
+  @Min(1)
   Integer id;
-  @NotBlank @Length(max = 255)
+  @NotBlank
+  @Length(max = 255)
   String name;
   @NotBlank @Length(max = 500)
   String description;
