@@ -1,11 +1,10 @@
 package at.kocmana.testservices.productservice.productreview.model.domain;
 
 import at.kocmana.testservices.productservice.productinformation.model.domain.Product;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import javax.persistence.*;
 
 @Entity
 @Table(name = "product_review")
