@@ -1,16 +1,16 @@
 package at.kocmana.testservices.productservice.productreview.model.dto;
 
+import jakarta.validation.constraints.Max;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotNull;
 import lombok.Value;
 import org.hibernate.validator.constraints.Length;
-
-import javax.validation.constraints.Max;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
 
 @Value
 public class ProductReviewUpdateRequest {
 
-  @NotNull @Min(1)
+  @NotNull
+  @Min(1)
   int id;
   @NotNull @Min(1)
   int customerId;
